Step1: 모델 분리
