+++ WeatherForecast/WeatherForecast/Source/Model/TempUnit.swift
@@ -1,0 +1,18 @@
+//
+//  TempUnit.swift
+//  WeatherForecast
+//
+//  Created by 홍승완 on 2024/03/12.
+//
+
+
+// MARK: - Temperature Unit
+enum TempUnit: String {
+    case metric, imperial
+    var expression: String {
+        switch self {
+        case .metric: return "℃"
+        case .imperial: return "℉"
+        }
+    }
+}
