+++ WeatherForecast/WeatherForecast/Source/Model/Weather.swift
@@ -6,14 +6,8 @@
 
 import Foundation
 
-// MARK: - Weather JSON Format
-class WeatherJSON: Decodable {
-    let weatherForecast: [WeatherForecastInfo]
-    let city: City
-}
-
 // MARK: - List
-class WeatherForecastInfo: Decodable {
+struct WeatherForecastInfo: Decodable {
     let dt: TimeInterval
     let main: MainInfo
     let weather: Weather
@@ -21,13 +15,13 @@ class WeatherForecastInfo: Decodable {
 }
 
 // MARK: - MainClass
-class MainInfo: Decodable {
+struct MainInfo: Decodable {
     let temp, feelsLike, tempMin, tempMax: Double
     let pressure, seaLevel, grndLevel, humidity, pop: Double
 }
 
 // MARK: - Weather
-class Weather: Decodable {
+struct Weather: Decodable {
     let id: Int
     let main: String
     let description: String
@@ -35,28 +29,36 @@ class Weather: Decodable {
 }
 
 // MARK: - City
-class City: Decodable {
+struct City: Decodable {
     let id: Int
     let name: String
     let coord: Coord
     let country: String
     let population, timezone: Int
     let sunrise, sunset: TimeInterval
+    
+    init(id: Int, name: String, coord: Coord, country: String, population: Int, timezone: Int, sunrise: TimeInterval, sunset: TimeInterval) {
+        self.id = id
+        self.name = name
+        self.coord = coord
+        self.country = country
+        self.population = population
+        self.timezone = timezone
+        self.sunrise = sunrise
+        self.sunset = sunset
+    }
+    
+    static let mock: City = City(id: 0, name: "", coord: Coord.mock, country: "", population: 0, timezone: 0, sunrise: 0, sunset: 0)
 }
 
 // MARK: - Coord
-class Coord: Decodable {
+struct Coord: Decodable {
     let lat, lon: Double
-}
-
-// MARK: - Temperature Unit
-enum TempUnit: String {
-    case metric, imperial
-    var expression: String {
-        switch self {
-        case .metric: return "℃"
-        case .imperial: return "℉"
-        }
+    
+    init(lat: Double, lon: Double) {
+        self.lat = lat
+        self.lon = lon
     }
+    
+    static let mock: Coord = Coord(lat: 0, lon: 0)
 }
-
