+++ WeatherForecast/WeatherForecast/Source/Scene/Weather/WeatherViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  WeatherViewModel.swift
+//  WeatherForecast
+//
+//  Created by 홍승완 on 2024/03/12.
+//
+
+import Foundation
